+++ src/gui/eventHandler/eventHandler.cpp
@@ -1,6 +1,6 @@
 #include "eventHandler.h"
 
-// TODO: Need to refactor so that the functions that take in "void* data" also take 
+// TODO: Need to refactor so that the functions that take in "void* data" also take
 // in "sf::Event eventData". As it is, "void* data" is being used to pass sf::Event
 // to the triggerEvent function and there is not a good way to pass any actual data
 // to the final action.
@@ -135,7 +135,7 @@ EventManager::EventManager()
  */
 EventManager::~EventManager()
 {
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         delete this->events[i];
     }
     CubeLog::info("Event manager destroyed");
@@ -162,7 +162,7 @@ int EventManager::createEvent(std::string eventName)
  */
 bool EventManager::removeEvent(EventHandler* event)
 {
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         if (this->events[i] == event) {
             delete this->events[i];
             this->events.erase(this->events.begin() + i);
@@ -186,7 +186,7 @@ bool EventManager::removeEvent(int index)
 
 bool EventManager::removeEvent(std::string eventName)
 {
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         if (this->events[i]->getName() == eventName) {
             delete this->events[i];
             this->events.erase(this->events.begin() + i);
@@ -225,7 +225,7 @@ std::vector<EventHandler*> EventManager::getEvents()
  */
 EventHandler* EventManager::getEvent(std::string eventName)
 {
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         if (this->events[i]->getName() == eventName) {
             return this->events[i];
         }
@@ -245,7 +245,7 @@ bool EventManager::triggerEvent(sf::Event event, void* data)
     if (checkClickableAreas(*((sf::Event*)data))) {
         return true;
     }
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         if (this->events[i]->getEventType() == event.type) {
             this->events[i]->triggerEvent(data);
             return true;
@@ -266,7 +266,7 @@ bool EventManager::triggerEvent(int index, void* data)
     if (checkClickableAreas(*((sf::Event*)data))) {
         return true;
     }
-    if(index >= this->events.size()){
+    if (index >= this->events.size()) {
         CubeLog::error("Index out of range");
         return false;
     }
@@ -286,7 +286,7 @@ bool EventManager::triggerEvent(std::string eventName, void* data)
     if (checkClickableAreas(*((sf::Event*)data))) {
         return true;
     }
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         if (this->events[i]->getName() == eventName) {
             this->events[i]->triggerEvent(data);
             return true;
@@ -307,7 +307,7 @@ bool EventManager::triggerEvent(sf::Event::EventType eventType, void* data)
     if (checkClickableAreas(*((sf::Event*)data))) {
         return true;
     }
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         if (this->events[i]->getEventType() == eventType && this->events[i]->getSpecificEventType() == SpecificEventTypes::NULL_EVENT) {
             this->events[i]->triggerEvent(data);
             return true;
@@ -328,7 +328,7 @@ bool EventManager::triggerEvent(SpecificEventTypes specificEventType, void* data
     // if(checkClickableAreas(*((sf::Event*)data))){
     //     return true;
     // }
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         if (this->events[i]->getSpecificEventType() == specificEventType && this->events[i]->getEventType() == sf::Event::Count) {
             this->events[i]->triggerEvent(data);
             return true;
@@ -342,7 +342,7 @@ bool EventManager::triggerEvent(SpecificEventTypes specificEventType, sf::Event:
     // if(checkClickableAreas(*((sf::Event*)data))){
     //     return true;
     // }
-    for (int i = 0; i < this->events.size(); i++) {
+    for (size_t i = 0; i < this->events.size(); i++) {
         if (this->events[i]->getEventType() == eventType && this->events[i]->getSpecificEventType() == specificEventType) {
             this->events[i]->triggerEvent(data);
             return true;
@@ -381,14 +381,16 @@ bool EventManager::checkClickableAreas(sf::Event event)
     if (distance < 5) {
         CubeLog::info("Distance is less than 5");
         for (ClickableArea* area : this->clickableAreas) {
-            if(!area->clickableObject->getIsClickable()) continue;
+            if (!area->clickableObject->getIsClickable())
+                continue;
             // get the x and y from the event
             int x = event.mouseButton.x;
             int y = event.mouseButton.y;
             // check if the x and y are within the area
             CubeLog::debugSilly("Checking clickable area: x:" + std::to_string(x) + " y:" + std::to_string(y) + " area: xMin:" + std::to_string(area->xMin) + " xMax:" + std::to_string(area->xMax) + " yMin:" + std::to_string(area->yMin) + " yMax:" + std::to_string(area->yMax));
             if (x < area->xMax && x > area->xMin && y < area->yMax && y > area->yMin) {
-                if(event.type == sf::Event::MouseButtonReleased) area->clickableObject->onRelease(&event);
+                if (event.type == sf::Event::MouseButtonReleased)
+                    area->clickableObject->onRelease(&event);
                 if (event.mouseButton.button == sf::Mouse::Left && event.type == sf::Event::MouseButtonReleased) {
                     area->clickableObject->onClick(&event);
                     return true;
@@ -397,30 +399,31 @@ bool EventManager::checkClickableAreas(sf::Event event)
                     area->clickableObject->onRightClick(&event);
                     return true;
                 }
-                if(event.type == sf::Event::MouseButtonPressed){
+                if (event.type == sf::Event::MouseButtonPressed) {
                     area->clickableObject->onMouseDown(&event);
                     return true;
                 }
             }
         }
-    }else if((sf::Mouse::isButtonPressed(sf::Mouse::Left) || sf::Mouse::isButtonPressed(sf::Mouse::Right) || sf::Touch::isDown(0))){
-        if(std::abs(float(yChange)) * (0.5f) > std::abs(float(xChange))){
+    } else if ((sf::Mouse::isButtonPressed(sf::Mouse::Left) || sf::Mouse::isButtonPressed(sf::Mouse::Right) || sf::Touch::isDown(0))) {
+        if (std::abs(float(yChange)) * (0.5f) > std::abs(float(xChange))) {
             // CubeLog::info("Dragged Y");
             sf::Event ev;
             ev.type = sf::Event::MouseWheelScrolled;
             ev.mouseWheelScroll.delta = -yChange;
             triggerEvent(SpecificEventTypes::DRAG_Y, &ev);
-        }else if(std::abs(float(xChange) * (0.5f)) > std::abs(float(yChange))){
+        } else if (std::abs(float(xChange) * (0.5f)) > std::abs(float(yChange))) {
             // CubeLog::info("Dragged X");
             sf::Event ev;
             ev.type = sf::Event::MouseWheelScrolled;
             ev.mouseWheelScroll.delta = xChange;
             triggerEvent(SpecificEventTypes::DRAG_X, &ev);
         }
-    }else{
-        for(ClickableArea* area : this->clickableAreas){
-            if(!area->clickableObject->getIsClickable()) continue;
-            if(event.type == sf::Event::MouseButtonReleased){
+    } else {
+        for (ClickableArea* area : this->clickableAreas) {
+            if (!area->clickableObject->getIsClickable())
+                continue;
+            if (event.type == sf::Event::MouseButtonReleased) {
                 area->clickableObject->onRelease(&event);
             }
         }
